Rationalise FilesenderRestClient.php, create Dads.class.php include file.
